chore: update dependencies and update to work with them (#96)
* chore: update dependencies and rewrite expected test results

I think babylon added a property to its ast structure for all nodes in a point release that messed up the tests for anyone trying to run them with a fresh `npm install`. This should fix it.

Fixes #95.

* chore: mark flow types as used/defined

* chore: test against node 4 & 6

* chore(package): fix repository type

* chore: update codebase for eslint v3

* chore: update codebase for magic-string 0.15.x

`insert` was replaced by `insertLeft` and `insertRight`. I think I got them all correct, but I still don't have an intuitive grasp of which to use when.