chore: fix build
There was a rollup error on travis that looks like it may have been due to
cached old dependencies; things worked locally for me. So update package.json to
explicitly specify those dependencies to bust the cache.

Also, there's now a new `identifierName` field in the babylon AST, so update all
tests to include those in the expected output.