fix: don't add a backslash to escape a character that's already escaped (#110)
Fixes https://github.com/decaffeinate/decaffeinate/issues/506

Really, we need to check if the number of preceding backshashes is odd, to
handle an arbitrary number of backslash characters before this one in the
string.

Also fix a subtle issue exposed by this change: in one code path, we do an
unescape followed by an escape on the same string, but the escape function is
given the original string instead of the new string. That means that in some
cases, we expect a transform like `\'` to `'` to `\'`. But since we use the
original string when escaping, it was hitting this new code path and not adding
the backslash. It would be nice if the code around this was more robust, but a
simple and correct-seeming fix is to avoid that case altogether when the
unescape and escape are for the same character.