fix: keep `var` declarations in loops in some unsafe-to-convert cases (#113)
Fixes https://github.com/decaffeinate/decaffeinate/issues/624

Loops make variables tricky because `let` and `const` declarations get a new
variable instance for each loop iteration. This introduces two important
semantic differences:
* Closures referencing `var` variables reference the same variable across all
  iterations, whereas for `let`/`const` variables they reference the variable
  from the current iteration.
* `var` declarations without an initializer will have the value from the
  previous iteration. `let` declarations without an initializer will start as
  `undefined`.

To address these cases, I made the `bindingCouldBeBlockScope` function more
conservative. When a variable declaration is within a loop (that exists in the
same function scope as the declaration), we keep `var` if any reference exists
in a different function scope (i.e. within a closure). In the remaining cases,
we only convert to `let` or `const` if the variable is a loop assignee, is
initialized when it's declared, or when it has a top-level assignment before any
usage. This misses some cases that could be detected by more advanced static
analysis, but this approach probalby gets all of the common cases.

While I was at it, I changed some babel usages to use more concise methods,
which I discovered by reading through the [Babel Handbook](https://github.com/thejameskyle/babel-handbook/).

Funny enough, after I fixed this, the problem still occurred in my codebase
because the eslint auto-fixer for the [no-var](http://eslint.org/docs/rules/no-var) rule has the same bug.
I'll follow up with the eslint folks on a fix for that.