fix: prevent block scope for init-referenced bindings
Previously we would turn e.g. `var x = x;` into a `let` (or `const`) declaration, but that introduces a TDZ violation since the initializer for the binding is inside the TDZ. We have to check that the initializer for the binding does not reference the binding being initialized.

I'm not entirely sure why babel-traverse does not count the `xxx` in `var-with-own-binding-in-initializer` as a `referencePath`, but it does show up in `constantViolations` so we use that too.

Closes decaffeinate/decaffeinate#724