fix: don't create octal literals when combining strings (#117)
Fixes https://github.com/decaffeinate/decaffeinate/issues/703

Almost every escape sequence will error if it is incomplete, so we don't need to
worry too much about a string having half of an escape sequence will be combined
later with another half. However, the one case that I'm aware of where this
could happen is `\0` (a zero byte) getting combined with a string starting with
a number, which creates an octal literal, which crashes in strict mode.