fix: don't convert `new` expressions to arrow functions
Progress toward https://github.com/decaffeinate/decaffeinate/issues/846

It's not allowed to call `new` on an arrow function, so if we can easily
determine that the function is in a `new` expression, skip it. Generally, this
property makes it impossible to determine for sure if it's safe to convert any
function to an arrow function, but this case seems obscure enough that it
shouldn't stop us from converting to arrow functions in normal-looking cases.