fix: properly remove the property when converting to destructure (#121)
Fixes https://github.com/decaffeinate/decaffeinate/issues/874

Previously, the code removed from the end of the object to the end of the
property, but if the object was surrounded in parens, which would break the
code. Now, we look through the tokens to find the dot token, and remove from
there to the end of the property.