fix: wrap arrow function bodies in parens in more cases (#122)
Fixes https://github.com/decaffeinate/decaffeinate/issues/791

The previous code would look at the AST to determine if the body expression
needs parens, but the actual case we need to handle is when the body is any
expression starting with a `{` token. So rather than look at the AST, we can
look at the code and see if the first character of the node is a `{` character,
which handles all previous cases plus some more complex cases that weren't
handled before.