fix: sync scope by registering new declarations
`@babel/traverse` has built-in `NodePath` manipulations like `insertBefore` and `replaceWith`, but these operations do not automatically discover new declarations in the inserted AST nodes. This change manually notifies the scope of new paths to find declarations in.

This prevents situations where a new variable declaration is added but not included in the scope, and a subsequent plugin goes looking for the scope binding associated with the new variable declaration and fails because the scope is out of sync.