fix: allow export declarations to be idiomatic
`@babel/traverse` has `Scope#generateUid` which generates a unique ID based on a given name. What I'd like for that method to do is to always use that name if possible, but to generate names based on it if not. That is not how it works, unfortunately, so I wrote my own based on it.

The upshot is that what was previously this:

```js
let _PI = 3.14;
export { _PI as PI };
```

is now this:

```js
export let PI = 3.14;
```

Much nicer.